pyproject.toml
[project]
name = "script-minio"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "minio>=7.2.16",
    "python-dotenv>=1.1.1",
]
